// webui/src/lib/gen/models/SolutionReferenceResponse.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { SolutionReferenceResponseName } from './SolutionReferenceResponseName';
import type { SolutionReferenceResponseSolutionIndex } from './SolutionReferenceResponseSolutionIndex';
import type { SolutionReferenceResponseObjectiveValues } from './SolutionReferenceResponseObjectiveValues';
import type { SolutionReferenceResponseVariableValues } from './SolutionReferenceResponseVariableValues';

/**
 * The response information provided when `SolutionReference` object are returned from the client.
 */
export interface SolutionReferenceResponse {
	name?: SolutionReferenceResponseName;
	solution_index: SolutionReferenceResponseSolutionIndex;
	state_id: number;
	objective_values: SolutionReferenceResponseObjectiveValues;
	variable_values: SolutionReferenceResponseVariableValues;
}