// webui/src/lib/gen/models/solutionReferenceResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { SolutionReferenceResponseName } from './solutionReferenceResponseName';
import type { SolutionReferenceResponseSolutionIndex } from './solutionReferenceResponseSolutionIndex';
import type { SolutionReferenceResponseObjectiveValues } from './solutionReferenceResponseObjectiveValues';
import type { SolutionReferenceResponseVariableValues } from './solutionReferenceResponseVariableValues';

/**
 * The response information provided when `SolutionReference` object are returned from the client.
 */
export interface SolutionReferenceResponse {
	name: SolutionReferenceResponseName;
	solution_index: SolutionReferenceResponseSolutionIndex;
	state_id: number;
	objective_values: SolutionReferenceResponseObjectiveValues;
	variable_values: SolutionReferenceResponseVariableValues;
}